Encode numbers as a single stream
Instead of one string per frame, use one giant string, with "`" as the
delimeter between frames.  This saves 1-2KB of ending and beginning
strings.

Instead of capping RLEs at 64, resulting in runs like [64,0,64,0,N],
introduce a three-byte encoding for numbers >= 64.  This saves another 1KB
or so, and simplifies both encoding and decoding logic.