Use multiple SFX and channels
Playing chords requires playing multiple notes at once, which isn't
possible if every guitar note reuses the same SFX slot.  So now SFX 0-3
are all guitar, and we round-robin among them.  That means we can play up
to four notes in a chord, which is the best the p8 can do anyway.

The wrong-note honk noise stops all playing notes.

The click track is gone, because it hogs a channel.  To bring it back,
use channels 0-2 instead; increment the channel counter mod-3 in `play`.