Fix song display behind score screen
The score screen is overlaid atop the final display of the song.  But now
that songs and the score screen use different display palettes, that
looked terrible: the notes and strings end up all sorts of weird colors.

The hack we use now is to draw the song behind the score using only shades
of grey.  `draw_song` now takes a theme parameter, which happily
centralizes all color selection in a single struct.  The result looks
pretty good and focuses attention better on the score dialog than what we
had before.